% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model_data}
\alias{model_data}
\title{Fit a statistical model to the data}
\usage{
model_data(data, formula, model_type = "lm")
}
\arguments{
\item{data}{A data frame to model}

\item{formula}{Model formula (e.g., `y ~ x1 + x2`)}

\item{model_type}{Type of model ("lm", "glm")}
}
\value{
A summary of the fitted model
}
\description{
Fit a statistical model to the data
}
