% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{visualize_data}
\alias{visualize_data}
\title{Generate customizable visualizations}
\usage{
visualize_data(data, type = "scatter", options = list())
}
\arguments{
\item{data}{A data frame to visualize}

\item{type}{Type of visualization ("scatter", "boxplot")}

\item{options}{A list of customization options (e.g., `x`, `y`, `color`, `fill`)}
}
\value{
A ggplot object
}
\description{
Generate customizable visualizations
}
